'amount' is now an integer

+++ database/database_implementation.sql
@@ -6,7 +6,7 @@ CREATE TABLE User (
 
 CREATE TABLE Debts (
   debtid integer AUTO_INCREMENT NOT NULL,
-  amount double NOT NULL,
+  amount integer NOT NULL,
   description varchar(200) NOT NULL,
   creditorid integer NOT NULL,
   debtorid integer NOT NULL,
